fix laser_duration casting

+++ Marlin/laser.h
@@ -27,7 +27,7 @@ typedef struct {
   float pwm; // PWM frequency used to control laser intensity in HZ
   float intensity; // Laser firing instensity 0.0 - 100.0
   float ppm; // pulses per millimeter, for pulsed firing mode
-  uint16_t duration; // laser firing duration in microseconds, for pulsed firing mode
+  unsigned long duration; // laser firing duration in microseconds, for pulsed firing mode
   bool status; // LASER_ON / LASER_OFF
   uint8_t mode; // LASER_CONTINUOUS, LASER_PULSED, LASER_RASTER
   unsigned long last_firing; // microseconds since last laser firing
